+++ packages/element/CHANGELOG.md
@@ -3,6 +3,86 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.4-beta.189](https://192.168.10.60/sc-dva-antd-pro/compare/@scboson/sc-element@1.0.4-beta.188...@scboson/sc-element@1.0.4-beta.189) (2023-12-28)
+
+**Note:** Version bump only for package @scboson/sc-element
+
+
+
+
+
+## [1.0.4-beta.188](https://192.168.10.60/sc-dva-antd-pro/compare/@scboson/sc-element@1.0.4-beta.187...@scboson/sc-element@1.0.4-beta.188) (2023-12-26)
+
+**Note:** Version bump only for package @scboson/sc-element
+
+
+
+
+
+## [1.0.4-beta.187](https://192.168.10.60/sc-dva-antd-pro/compare/@scboson/sc-element@1.0.4-beta.186...@scboson/sc-element@1.0.4-beta.187) (2023-12-25)
+
+**Note:** Version bump only for package @scboson/sc-element
+
+
+
+
+
+## [1.0.4-beta.186](https://192.168.10.60/sc-dva-antd-pro/compare/@scboson/sc-element@1.0.4-beta.185...@scboson/sc-element@1.0.4-beta.186) (2023-12-22)
+
+**Note:** Version bump only for package @scboson/sc-element
+
+
+
+
+
+## [1.0.4-beta.185](https://192.168.10.60/sc-dva-antd-pro/compare/@scboson/sc-element@1.0.4-beta.184...@scboson/sc-element@1.0.4-beta.185) (2023-12-13)
+
+**Note:** Version bump only for package @scboson/sc-element
+
+
+
+
+
+## [1.0.4-beta.184](https://192.168.10.60/sc-dva-antd-pro/compare/@scboson/sc-element@1.0.4-beta.183...@scboson/sc-element@1.0.4-beta.184) (2023-12-07)
+
+**Note:** Version bump only for package @scboson/sc-element
+
+
+
+
+
+## [1.0.4-beta.183](https://192.168.10.60/sc-dva-antd-pro/compare/@scboson/sc-element@1.0.4-beta.182...@scboson/sc-element@1.0.4-beta.183) (2023-11-14)
+
+**Note:** Version bump only for package @scboson/sc-element
+
+
+
+
+
+## [1.0.4-beta.182](https://192.168.10.60/sc-dva-antd-pro/compare/@scboson/sc-element@1.0.4-beta.181...@scboson/sc-element@1.0.4-beta.182) (2023-11-13)
+
+**Note:** Version bump only for package @scboson/sc-element
+
+
+
+
+
+## [1.0.4-beta.181](https://192.168.10.60/sc-dva-antd-pro/compare/@scboson/sc-element@1.0.4-beta.180...@scboson/sc-element@1.0.4-beta.181) (2023-09-01)
+
+**Note:** Version bump only for package @scboson/sc-element
+
+
+
+
+
+## [1.0.4-beta.180](https://192.168.10.60/sc-dva-antd-pro/compare/@scboson/sc-element@1.0.4-beta.179...@scboson/sc-element@1.0.4-beta.180) (2023-09-01)
+
+**Note:** Version bump only for package @scboson/sc-element
+
+
+
+
+
 ## [1.0.4-beta.179](https://192.168.10.60/sc-dva-antd-pro/compare/@scboson/sc-element@1.0.4-beta.178...@scboson/sc-element@1.0.4-beta.179) (2023-08-10)
 
 **Note:** Version bump only for package @scboson/sc-element